Revert "CASMCMS-8505: Support for multi-architecture aarch64 builds"
This reverts commit 387d2db828e147be4db44e5ed50c602a64755b7c.
