+++ kubernetes/cms-ipxe/templates/configmap-settings.yaml
@@ -29,20 +29,12 @@ metadata:
     app.kubernetes.io/name: {{ .Values.ipxe.name }}
 data:
   settings.yaml: |
-    # Global build options
-    cray_ipxe_token_host: {{ .Values.ipxe.api_gw }}
-    cray_ipxe_token_min_remaining_valid_time: {{ .Values.ipxe.token_min_remaining_valid_time }}
-    cray_ipxe_build_with_cert: {{ .Values.ipxe.build_with_cert }}
-    cray_ipxe_build_service_log_level : {{ .Values.ipxe.build_service_log_level }}
-
-    # x86_64 build options
     cray_ipxe_build_x86: {{ .Values.ipxe.build_x86 }}
+    cray_ipxe_build_bss: {{ .Values.ipxe.build_bss }}
+    cray_ipxe_build_shell: {{ .Values.ipxe.build_shell }}
+    cray_ipxe_build_with_cert: {{ .Values.ipxe.build_with_cert }}
+    cray_ipxe_token_host: {{ .Values.ipxe.api_gw }}
     cray_ipxe_binary_name: {{ .Values.ipxe.cray_ipxe_binary_name }}
-    cray_ipxe_debug_enabled: {{ .Values.ipxe.build_x86_debug }}
     cray_ipxe_debug_binary_name: {{ .Values.ipxe.cray_ipxe_debug_binary_name }}
-
-    # aarch64 build options
-    cray_ipxe_aarch64_enabled: {{ .Values.ipxe.build_aarch64 }}
-    cray_ipxe_aarch64_binary_name: {{ .Values.ipxe.cray_ipxe_aarch64_binary_name }}
-    cray_ipxe_aarch64_debug_enabled : {{ .Values.ipxe.build_aarch64_debug }}
-    cray_ipxe_aarch64_debug_binary_name: {{ .Values.ipxe.cray_ipxe_aarch64_debug_binary_name }}
+    cray_ipxe_build_arch: {{  .Values.ipxe.cray_ipxe_build_arch }}
+    cray_ipxe_build_kind: {{  .Values.ipxe.cray_ipxe_build_kind }}
