+++ src/crayipxe/__init__.py
@@ -21,26 +21,3 @@
 # ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 # OTHER DEALINGS IN THE SOFTWARE.
 #
-
-import os
-import logging
-
-LOGGER = logging.getLogger(__name__)
-IPXE_BUILD_DIR = '/ipxe'
-
-# Format logs and set the requested log level.
-log_format = "%(asctime)-15s - %(levelname)-7s - %(name)s - %(message)s"
-requested_log_level = os.environ.get('LOG_LEVEL', 'INFO')
-log_level = logging.getLevelName(requested_log_level)
-
-bad_log_level = None
-if type(log_level) != int:
-    bad_log_level = requested_log_level
-    log_level = logging.INFO
-
-logging.basicConfig(level=log_level, format=log_format)
-if bad_log_level:
-    LOGGER.warning('Log level %r is not valid. Falling back to INFO',
-                   bad_log_level)
-
-
