+++ src/crayipxe/liveness/ipxe_timestamp.py
@@ -1,7 +1,7 @@
 #
 # MIT License
 #
-# (C) Copyright 2020-2023 Hewlett Packard Enterprise Development LP
+# (C) Copyright 2020-2022 Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a
 # copy of this software and associated documentation files (the "Software"),
@@ -31,11 +31,10 @@ import datetime
 import json
 import logging
 import os
-import threading
-import time
 
 LOGGER = logging.getLogger(__name__)
-LIVENESS_PATH = '/tmp/ipxe_build_in_progress'
+IPXE_PATH='/tmp/ipxe_build_in_progress'
+DEBUG_IPXE_PATH='/tmp/debug_ipxe_build_in_progress'
 
 
 class BaseipxeTimestampException(BaseException):
@@ -71,15 +70,21 @@ class ipxeTimestamp(object):
             pass
 
         if not when:
-            timestamp = datetime.datetime.now().timestamp()
+            self.timestamp = datetime.datetime.now().timestamp()
         else:
-            timestamp = when.timestamp()
+            self.timestamp = when.timestamp()
+
+        self.expiration = self.timestamp + float(max_age)
 
         with open(self.path, 'w') as timestamp_file:
-            data = {'timestamp': float(timestamp),
-                    'max_age': float(max_age),
-                    'last_build': None}
+            data = {'timestamp': self.timestamp,
+                    'expiration': self.expiration}
+            LOGGER.debug("Created timestamp file: %s "
+                         "-- Timestamp: %s "
+                         "Expiration: %s", path, data['timestamp'],
+                         data['expiration'])
             json.dump(data, timestamp_file)
+            
 
     @classmethod
     def byref(cls, path):
@@ -99,198 +104,45 @@ class ipxeTimestamp(object):
         if os.path.exists(path):
             self = cls.__new__(cls)
             self.path = path
+            with open(self.path, 'r') as timestamp_file:
+                data = json.load(timestamp_file)
+                self.timestamp = float(data['timestamp'])
+                self.expiration = float(data['expiration'])
+                LOGGER.debug("Opened timestamp file: %s "
+                             "-- Timestamp: %s "
+                             "Expiration: %s", path, data['timestamp'],
+                             data['expiration'])
             return self
         else:
             raise ipxeTimestampNoEnt
 
     @property
-    def _data(self):
+    def alive(self):
         """
-        A read-only, non-caching method for parsing information from disk.
-        :return: A dictionary from the parsed timestamp file.
-        """
-        if os.path.exists(self.path):
-            with open(self.path, 'r') as timestamp_file:
-                data = json.load(timestamp_file)
-                return data
-        else:
-            # Newly instantiated objects will have no persistent data on disk.
-            return {'timestamp': None, 'max_age': None, 'last_build': None}
+        Has the time stamp expired?
 
-    @_data.setter
-    def _data(self, dict_obj):
+        Return
+          True -- The timestamp has not expired
+          False -- The timestamp has expired
         """
-        :param dict_obj: A dictionary to be written to disk.
-        :return: None.
-        """
-        with open(self.path, 'w') as timestamp_file:
-            json.dump(dict_obj, timestamp_file)
+        return datetime.datetime.now() < datetime.datetime.fromtimestamp(self.expiration)
 
     @property
     def value(self):
         """
-        The float value of when the timestamp was last pressed to disk.
-        :return: A float value as stored in a file.
+        The timestamp value, as stored on disk. This property does not cache
+        the value; instead it reads it each time the property is accessed.
         """
         try:
-            return datetime.datetime.fromtimestamp(float(self._data['timestamp']))
-        except KeyError:
-            return None
-
-    @value.setter
-    def value(self, when=None):
-        """
-        :param when: A float timestamp or nothing at all
-        :return: None.
-        """
-        if not when:
-            # How?
-            when = datetime.datetime.now().timestamp()
-        data = self._data
-        data['timestamp'] = when
-        self._data = data
-
-    @property
-    def value_age(self):
-        return datetime.datetime.now() - self.value
-
-    @property
-    def last_build(self):
-        """
-        :return: A datetime value representing the last time the deployment was refreshed.
-        """
-        try:
-            return datetime.datetime.fromtimestamp(float(self._data['last_build']))
-        except (KeyError, TypeError):
-            return None
-
-    @last_build.setter
-    def last_build(self, when=None):
-        """
-        :param when: a float timestamp representing when the last build was a success, or None, representing right now.
-        :return: None
-        """
-        if not when:
-            # How?
-            when = datetime.datetime.now().timestamp()
-        data = self._data
-        data['last_build'] = when
-        self._data = data
-
-    @property
-    def last_build_age(self):
-        """
-        :return: a timedelta object showing how long ago a build was made, or 'Never' if the build has never been
-        completed.
-        """
-        last_build = self.last_build
-        if not last_build:
-            return 'Never'
-        else:
-            return datetime.datetime.now() - self.last_build
-
-    @property
-    def max_age(self):
-        """
-        The maximum age that a timestamp is considered valid in the form of a float representing seconds.
-        :return: a timedelta value
-        """
-        return float(self._data['max_age'])
-
-    @max_age.setter
-    def max_age(self, val):
-        """
-        Sets a new maximum age for the heartbeat/build value (in seconds)
-        :param val: a float value (in seconds)
-        :return: None
-        """
-        data = self._data
-        data['max_age'] = val
-        self._data = data
-
-    @property
-    def max_delta(self):
-        """
-        A representation of the amount of time a timestamp can be considered valid based on a timedelta interpretation.
-        :return: a timedelta object instance
-        """
-        return datetime.timedelta(seconds=self.max_age)
-
-    @property
-    def expiration_date(self):
-        """
-        The datetime value when the timestamp will next expire.
-        :return: a datetime.datetime object
-        """
-        return self.value + self.max_delta
-
-    @property
-    def expired(self):
-        """
-        A Boolean value indicating if a timestamp has expired based on thread update from self.value.
-        :return: A bool value indicating overall expiration of the thread timestamp
-        """
-        return datetime.datetime.now() > self.expiration_date
-
-    @property
-    def recent_build(self):
-        """
-        :return: Boolean value indicating that we have had a recent build.
-        """
-        last_build_age = self.last_build_age
-        if not isinstance(last_build_age, datetime.timedelta):
-            return False
-        return self.last_build_age < self.max_delta
-
-    @property
-    def alive(self):
-        """
-        Return
-          True -- The timestamp has not expired or we have a recent build
-          False -- The timestamp has expired and no evidence of a recent build
-        """
-        # Any recent signs of life
-        return any([not self.expired,
-                    self.recent_build])
-
-    def refresh(self):
-        """
-        Update the timestamp with the current time, giving the timestamp object a new lease on life. Does not modify
-        existing max_age values set.
-        :return: None
-        """
-        self.value = float(datetime.datetime.now().timestamp())
-
-    def refresh_build(self):
-        """
-        Signals to the builder that a new build has just been created.
-        :return: None
-        """
-        self.last_build = None
+            with open(self.path, 'r') as timestamp_file:
+                data = json.load(timestamp_file)
+                return datetime.datetime.fromtimestamp(float(data['timestamp']))
+        except FileNotFoundError:
+            LOGGER.warning("Timestamp never initialized to '%s'" % (self.path))
+            return datetime.datetime.fromtimestamp(0)
 
     def delete(self):
         """
         Delete the timestamp file
         """
         os.remove(self.path)
-
-    def __repr__(self):
-        buffer = ['Timestamp (%s):' % self.path]
-        buffer.append('\tLast Touched: %s - Age: %s' % (self.value, self.value_age))
-        buffer.append('\tLast Build: %s - Age: %s' % (self.last_build, self.last_build_age))
-        buffer.append('\tExpires: %s' % self.expiration_date)
-        buffer.append('\tAlive: %s' % (bool(self.alive)))
-        return '\n'.join(buffer)
-
-
-def liveliness_heartbeat(path):
-    """
-    Periodically add a timestamp to disk; this allows for reporting of basic
-    health at a minimum rate. This prevents the pod being marked as dead if
-    a period of no events have been monitored from k8s for an extended
-    period of time.
-    """
-    timestamp = ipxeTimestamp.byref(path)
-    while True:
-        timestamp.refresh()
-        time.sleep(10)
