+++ update_external_versions.conf
@@ -79,8 +79,8 @@
 # no images will be found.
 
 image: cray-tpsw-ipxe
-    major: 3
-    minor: 0
+    major: 2
+    minor: 4
     outfile: .cray-tpsw-ipxe-version
     server: algol60
     source: docker
